Next piece in the scoreboard

+++ projetC++/header/drawing.hpp
@@ -3,7 +3,7 @@
 
 #include "tetromino.hpp"
 
-void drawPixel(int x, int y, sf::Color color, sf::RenderWindow &window);
+void drawPixel(float x, float y, sf::Color color, sf::RenderWindow &window);
 void drawGrid(sf::RenderWindow& window);
 void drawLocked(Board& board,sf::RenderWindow& window);
 void drawScoreBoard(sf::RenderWindow& window,int lineCounter,int score);
